% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_enviro_data}
\alias{all_enviro_data}
\title{All Environmental Data}
\format{
An object of class \code{data.frame} with 1323793 rows and 10 columns.
}
\usage{
all_enviro_data
}
\description{
All Environmental Data
}
\keyword{datasets}
