% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_fish_data}
\alias{all_fish_data}
\title{All Fish Data}
\format{
An object of class \code{data.frame} with 2296 rows and 10 columns.
}
\usage{
all_fish_data
}
\description{
All Fish Data
}
\keyword{datasets}
