% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_prey_density}
\alias{all_prey_density}
\title{Aggregated Prey Density Data from Multiple Authors and Across Watersheds}
\format{
a dataframe
}
\source{
something
}
\usage{
all_prey_density
}
\description{
Aggregated Prey Density Data from Multiple Authors and Across Watersheds
}
\details{
Datasets included in final dataset are:

\strong{Corline} \cr
Something about Corline

\strong{Cordoleani} \cr
Something

\strong{Zeug} \cr
Something

\strong{Montgomery} \cr

\strong{Zooper R library} \cr
Something
\href{https://github.com/InteragencyEcologicalProgram/zooper}

\strong{Guignard} \cr
Something
}
\keyword{datasets}
