% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cordoleani_prey_data}
\alias{cordoleani_prey_data}
\title{Flora Cordoleani Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2779 rows and 12 columns.
}
\usage{
cordoleani_prey_data
}
\description{
Flora Cordoleani Data
}
\keyword{datasets}
