% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corline_prey_data}
\alias{corline_prey_data}
\title{Nicholas Corline Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2270 rows and 13 columns.
}
\usage{
corline_prey_data
}
\description{
Nicholas Corline Data
}
\keyword{datasets}
