% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{montgomery_prey_data}
\alias{montgomery_prey_data}
\title{Jacob Montgomery Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9353 rows and 12 columns.
}
\usage{
montgomery_prey_data
}
\description{
Jacob Montgomery Data
}
\keyword{datasets}
