% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthly_fish_data}
\alias{monthly_fish_data}
\title{Monthly Fish Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34 rows and 10 columns.
}
\usage{
monthly_fish_data
}
\description{
Monthly Fish Data
}
\keyword{datasets}
