% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zeug_merced_prey_data}
\alias{zeug_merced_prey_data}
\title{Steve Zeug Merced Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 140 rows and 11 columns.
}
\usage{
zeug_merced_prey_data
}
\description{
Steve Zeug Merced Data
}
\keyword{datasets}
