% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guignard_prey_data}
\alias{guignard_prey_data}
\title{Jason Guignard Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 156 rows and 12 columns.
}
\usage{
guignard_prey_data
}
\description{
Jason Guignard Data
}
\keyword{datasets}
