% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zeug_merced_prey_data}
\alias{zeug_merced_prey_data}
\title{Steve Zeug Merced Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 140 rows and 11 columns.
}
\usage{
zeug_merced_prey_data
}
\description{
Zeug provided prey density data from two rivers: San Joaquin River and Merced River.
}
\details{
San Joaquin River: \cr
Zeug collected data in 2016 for four sites along the San Joaquin River.

Merced River: \cr
Zeug collected data in April 2019 for four sites along the Merced River.

Citation: TODO
}
\keyword{datasets}
