% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zeug_sanJoaquin_prey_data}
\alias{zeug_sanJoaquin_prey_data}
\title{Steve Zeug San Joaquin Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55 rows and 13 columns.
}
\usage{
zeug_sanJoaquin_prey_data
}
\description{
Steve Zeug San Joaquin Data
}
\keyword{datasets}
