% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zooper_prey_data}
\alias{zooper_prey_data}
\title{Zooper Library Data}
\format{
An object of class \code{data.frame} with 596828 rows and 11 columns.
}
\usage{
zooper_prey_data
}
\description{
Zooper Library Data
}
\keyword{datasets}
